feat(calendar): integrate RoomService to dynamically load operating rooms and enhance schedule loading logic using Python algorithm

+++ or-scheduler-core/src/main/java/octacare/orschedulercore/dto/schedule/ScheduleGenerationResponse.java
@@ -7,8 +7,8 @@ import com.fasterxml.jackson.databind.annotation.JsonNaming;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public record ScheduleGenerationResponse (
         List<ScheduledSurgeryDto> schedule,
-        Integer score,
+        Double score,
         List<String> conflicts,
-        Integer generationTimeMs,
+        Long generationTimeMs,
         String status
 ) {}
